move ppo sample reraveling functions into training
