+++ atari_irl/training.py
@@ -7,7 +7,7 @@ from baselines.ppo2 import ppo2
 
 from rllab.sampler.base import BaseSampler
 
-from . import policies
+from . import policies, utils
 
 from collections import deque, namedtuple
 import time
@@ -164,8 +164,38 @@ class Runner(ppo2.AbstractEnvRunner, BaseSampler):
         self.lam = lam
         self.gamma = gamma
 
-    def set_algo(self, algo):
-        self.algo = algo
+    @staticmethod
+    def invert_ppo_sample_raveling(ppo_samples, num_envs=8):
+        return ppo_samples.reshape(
+            num_envs, ppo_samples.shape[0] // num_envs,
+            *ppo_samples.shape[1:]
+        ).swapaxes(1, 0)
+
+    @staticmethod
+    def ppo_samples_to_trajectory_format(ppo_samples, num_envs=8):
+        OBS_IDX = 0
+        ACTS_IDX = 3
+
+        obs = Runner.invert_ppo_sample_raveling(
+            ppo_samples[OBS_IDX], num_envs=num_envs
+        )
+        acts = Runner.invert_ppo_sample_raveling(
+            ppo_samples[ACTS_IDX], num_envs=num_envs
+        )
+        T = obs.shape[0]
+        observations = [[] for _ in range(num_envs)]
+        actions = [[] for _ in range(num_envs)]
+        assert acts.shape[0] == T
+        for t in range(T):
+            for i, (o, a) in enumerate(zip(obs[t], acts[t])):
+                observations[i].append(o)
+                actions[i].append(a)
+        trajectories = [{
+                            'observations': np.array(observations[i]),
+                            'actions': utils.one_hot(actions[i], 6)
+                        } for i in range(num_envs)]
+        np.random.shuffle(trajectories)
+        return trajectories
 
     def sample(self):
         mb_obs, mb_rewards, mb_actions, mb_values, mb_dones, mb_neglogpacs = [],[],[],[],[],[]
